+++ R/objective.R
@@ -189,11 +189,11 @@ Objective <- R6::R6Class(
         check = oeli::check_missing(argument_name),
         var_name = "argument_name"
       )
-      private$.check_argument_specified(argument_name, .verbose = .verbose)
       oeli::input_check_response(
         check = checkmate::check_flag(.verbose),
         var_name = ".verbose"
       )
+      private$.check_argument_specified(argument_name, .verbose = .verbose)
 
       ### get argument
       if (.verbose) {
@@ -213,11 +213,11 @@ Objective <- R6::R6Class(
         check = oeli::check_missing(argument_name),
         var_name = "argument_name"
       )
-      private$.check_argument_specified(argument_name, .verbose = .verbose)
       oeli::input_check_response(
         check = checkmate::check_flag(.verbose),
         var_name = ".verbose"
       )
+      private$.check_argument_specified(argument_name, .verbose = .verbose)
 
       ### remove argument
       if (.verbose) {
@@ -444,9 +444,10 @@ Objective <- R6::R6Class(
         )
         private$.gradient$evaluate(.at = .at, .negate = .negate, ...)
       } else {
-        cli::cli_abort(
-          "Gradient function is required but not specified, please call
-          {.var $set_gradient()} first.",
+        cli::cli_abort(paste(
+            "Gradient function is required but not specified, please call",
+            "{.var $set_gradient()} first."
+          ),
           call = NULL
         )
       }
@@ -463,9 +464,10 @@ Objective <- R6::R6Class(
         )
         private$.hessian$evaluate(.at = .at, .negate = .negate, ...)
       } else {
-        cli::cli_abort(
-          "Hessian function is required but not specified, please call
-          {.var $set_hessian()} first.",
+        cli::cli_abort(paste(
+            "Hessian function is required but not specified, please call",
+            "{.var $set_hessian()} first."
+          ),
           call = NULL
         )
       }
@@ -571,7 +573,8 @@ Objective <- R6::R6Class(
         return(private$.verbose)
       } else {
         oeli::input_check_response(
-          check = checkmate::check_flag(value)
+          check = checkmate::check_flag(value),
+          var_name = "verbose"
         )
         private$.verbose <- value
       }
@@ -644,36 +647,53 @@ Objective <- R6::R6Class(
     .arguments = list(),
     .seconds = Inf,
     .hide_warnings = FALSE,
-    .verbose = TRUE,
+    .verbose = getOption("verbose", default = FALSE),
     .gradient = NULL,
     .hessian = NULL,
 
     ### helper function that checks the target argument
-    .check_target = function(.at) {
-      if (!checkmate::test_numeric(
-        .at, any.missing = FALSE, len = sum(private$.npar)
-      )) {
-        variable_name <- oeli::variable_name(.at, fallback = ".at")
-        cli::cli_abort(
-          "Input {.var {variable_name}} must be a {.cls numeric} of length
-          {sum(private$.npar)}.",
-          call = NULL
-        )
+    .check_target = function(.at, .verbose = self$verbose) {
+
+      oeli::input_check_response(
+        check = checkmate::check_numeric(
+          .at, any.missing = FALSE, len = sum(private$.npar)
+        ),
+        var_name = oeli::variable_name(.at, fallback = ".at")
+      )
+      oeli::input_check_response(
+        check = checkmate::check_flag(.verbose),
+        var_name = ".verbose"
+      )
+      if (.verbose) {
+        cli::cli_alert_success(paste(
+          "The value{?s} for the {length(private$.npar)} target argument{?s}",
+          "{?is/are} correctly specified."
+        ))
       }
       invisible(TRUE)
+
     },
 
     ### helper function that checks if a function argument is specified
-    .check_argument_specified = function(argument_name, .verbose = self$verbose) {
+    .check_argument_specified = function(
+      argument_name, .verbose = self$verbose
+    ) {
+
       oeli::input_check_response(
         check = checkmate::check_string(argument_name),
         var_name = "argument_name"
       )
+      oeli::input_check_response(
+        check = checkmate::check_flag(.verbose),
+        var_name = ".verbose"
+      )
+
       if (!argument_name %in% names(private$.arguments)) {
-        cli::cli_abort(
-          "Function argument {.var {argument_name}} is required but not
-          specified, please call
-          {.var $set_argument({.val {argument_name}} = ...)} first.",
+        cli::cli_abort(paste(
+            "Function argument {.var {argument_name}} is required but not",
+            "specified, please call",
+            "{.var $set_argument({.val {argument_name}} = ...)} first."
+          ),
           call = NULL
         )
       }
@@ -682,6 +702,29 @@ Objective <- R6::R6Class(
           "Required argument {.val {argument_name}} is specified."
         )
       }
+      invisible(TRUE)
+
+    },
+
+    ### helper function that checks if all required arguments are specified
+    .check_arguments_complete = function(.verbose = self$verbose) {
+
+      oeli::input_check_response(
+        check = checkmate::check_flag(.verbose),
+        var_name = ".verbose"
+      )
+      arguments_required <- oeli::function_arguments(
+        private$.f, with_default = FALSE, with_ellipsis = FALSE
+      )
+      for (argument_name in setdiff(arguments_required, private$.target)) {
+        private$.check_argument_specified(argument_name, .verbose = FALSE)
+      }
+      if (.verbose) {
+        cli::cli_alert_success(
+          "All required fixed arguments are specified."
+        )
+      }
+
     },
 
     ### helper function to synchronize arguments with gradient and Hessian